+++ src/Hacks/Universal/ShowTouches/TouchTrail.cpp
@@ -25,29 +25,39 @@ bool CCTouchTrail::init(CCTouch* touch)
 
     clip->addChild(streak);
 
+    durationLabel = CCLabelBMFont::create("", "bigFont.fnt");
+
     update(0);
 
     this->addChild(clip);
     this->addChild(circle);
+    this->addChild(durationLabel);
     return true;
 }
 
 void CCTouchTrail::update(float dt)
 {
     dt = Speedhack::get()->getRealDeltaTime();
+    duration += dt;
 
     auto pos = touch->getLocation();
 
     streak->setPosition(pos);
     circle->setPosition(pos);
+    durationLabel->setPosition(pos);
 
     circle->setColor(CircleColour::get()->getColour());
     streak->setColor(TrailColour::get()->getColour());
 
-    circle->setScale(TrailScale::get()->getStringFloat() * 0.2f);
-    streak->setStroke(TrailScale::get()->getStringFloat() * 10);
+    circle->setScale(CircleScale::get()->getStringFloat() * 0.2f);
+    streak->setStroke(CircleScale::get()->getStringFloat() * 10);
+    durationLabel->setScale(CircleScale::get()->getStringFloat() * 0.3f);
+
+    streak->setVisible(TrailEnabled::get()->getRealEnabled());
+    durationLabel->setVisible(ShowTouchDuration::get()->getRealEnabled());
 
     streak->CCMotionStreak::update(dt);
+    durationLabel->setString(fmt::format("{:.2f}", duration).c_str());
 }
 
 CCTouchTrail* CCTouchTrail::create(CCTouch* touch)