+++ src/Labels/LabelConfig.cpp
@@ -42,6 +42,11 @@ LabelConfig LabelConfig::createFromObject(matjson::Value obj)
     if (obj.contains("label_type") && obj["label_type"].isNumber())
         conf.type = (LabelType)obj["label_type"].asInt().unwrap();
 
+    if (obj.contains("image_location") && obj["image_location"].isString())
+        conf.imageLocation = obj["image_location"].asString().unwrap();
+
+    
+
     if (obj.contains("events") && obj["events"].isArray())
     {
         for (auto obj : obj["events"].asArray().unwrap())
@@ -72,6 +77,7 @@ matjson::Value LabelConfig::save()
     obj["noclip_only"] = noclipOnly;
     obj["visible"] = visible;
     obj["label_type"] = (int)type;
+    obj["image_location"] = imageLocation;
 
     matjson::Value eventsArr = obj.array();