+++ src/Labels/Nodes/ImageNode.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "LabelNode.hpp"
+
+class ImageNode : public LabelNode
+{
+    protected:
+        CCSprite* sprite = nullptr;
+
+    public:
+        CREATE_FUNC(ImageNode);
+
+        virtual void labelConfigUpdated();
+        virtual void update(float dt);
+};